+++ test/data.js
@@ -4,7 +4,7 @@ import delay from 'delay';
 import Alpha from '../';
 const alpha = Alpha({ key: process.env.AV_KEY });
 
-jasmine.DEFAULT_TIMEOUT_INTERVAL = 30000;
+jest.setTimeout(30000);
 jest.unmock('cross-fetch');
 const TIME = 1000;
 
@@ -12,7 +12,7 @@ test(`intraday data works`, () => {
   expect.assertions(2);
   return delay(TIME)
     .then(() => alpha.data.intraday(`msft`))
-    .then(data => {
+    .then((data) => {
       expect(data['Meta Data']).toBeDefined();
       expect(data['Time Series (1min)']).toBeDefined();
     });
@@ -22,7 +22,7 @@ test(`daily data works`, () => {
   expect.assertions(2);
   return delay(TIME)
     .then(() => alpha.data.daily(`msft`))
-    .then(data => {
+    .then((data) => {
       expect(data['Meta Data']).toBeDefined();
       expect(data['Time Series (Daily)']).toBeDefined();
     });
@@ -32,7 +32,7 @@ test(`daily adjusted data works`, () => {
   expect.assertions(2);
   return delay(TIME)
     .then(() => alpha.data.daily_adjusted(`msft`))
-    .then(data => {
+    .then((data) => {
       expect(data['Meta Data']).toBeDefined();
       expect(data['Time Series (Daily)']).toBeDefined();
     });
@@ -42,7 +42,7 @@ test(`weekly data works`, () => {
   expect.assertions(2);
   return delay(TIME)
     .then(() => alpha.data.weekly(`msft`))
-    .then(data => {
+    .then((data) => {
       expect(data['Meta Data']).toBeDefined();
       expect(data['Weekly Time Series']).toBeDefined();
     });
@@ -52,7 +52,7 @@ test(`weekly adjusted data works`, () => {
   expect.assertions(2);
   return delay(TIME)
     .then(() => alpha.data.weekly_adjusted(`msft`))
-    .then(data => {
+    .then((data) => {
       expect(data['Meta Data']).toBeDefined();
       expect(data['Weekly Adjusted Time Series']).toBeDefined();
     });
@@ -62,7 +62,7 @@ test(`monthly data works`, () => {
   expect.assertions(2);
   return delay(TIME)
     .then(() => alpha.data.monthly(`msft`))
-    .then(data => {
+    .then((data) => {
       expect(data['Meta Data']).toBeDefined();
       expect(data['Monthly Time Series']).toBeDefined();
     });
@@ -72,7 +72,7 @@ test(`weekly adjusted data works`, () => {
   expect.assertions(2);
   return delay(TIME)
     .then(() => alpha.data.monthly_adjusted(`msft`))
-    .then(data => {
+    .then((data) => {
       expect(data['Meta Data']).toBeDefined();
       expect(data['Monthly Adjusted Time Series']).toBeDefined();
     });
@@ -82,7 +82,7 @@ test(`global quote data works`, () => {
   expect.assertions(2);
   return delay(TIME)
     .then(() => alpha.data.quote(`msft`))
-    .then(data => {
+    .then((data) => {
       expect(data['Global Quote']).toBeDefined();
       expect(data['Global Quote']['01. symbol']).toEqual('MSFT');
     });
@@ -92,7 +92,7 @@ test(`symbol search works`, () => {
   expect.assertions(2);
   return delay(TIME)
     .then(() => alpha.data.search(`Advanced Micro`))
-    .then(data => {
+    .then((data) => {
       expect(data['bestMatches']).toBeDefined();
       expect(data['bestMatches'][0]['1. symbol']).toEqual('AMD');
     });